pyproject.toml
[tool.poetry]
name = "coding-agents"
version = "0.1.0"
description = "AI Agents for GitHub SDLC Automation"
authors = ["Admin <admin@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.0"
openai = "^1.0"
pygithub = "^2.0"
gitpython = "^3.1"
tenacity = "^8.0"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.3"
mypy = "^1.8"
black = "^24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.mypy]
python_version = "3.11"
strict = true
